+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/headervalidationrules/BlobGasValidationRule.java
@@ -17,6 +17,7 @@ package org.hyperledger.besu.ethereum.mainnet.headervalidationrules;
 import org.hyperledger.besu.datatypes.BlobGas;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
 import org.hyperledger.besu.ethereum.mainnet.DetachedBlockHeaderValidationRule;
+import org.hyperledger.besu.evm.gascalculator.CancunGasCalculator;
 import org.hyperledger.besu.evm.gascalculator.GasCalculator;
 
 import org.slf4j.Logger;
@@ -35,7 +36,7 @@ public class BlobGasValidationRule implements DetachedBlockHeaderValidationRule 
 
   /**
    * Validates the block header by checking if the header's excess blob gas matches the calculated
-   * value based on the parent header.
+   * value based on the parent header, as well that the used blobGas is a multiple of GAS_PER_BLOB.
    */
   @Override
   public boolean validate(final BlockHeader header, final BlockHeader parent) {
@@ -51,6 +52,13 @@ public class BlobGasValidationRule implements DetachedBlockHeaderValidationRule 
           "Invalid block header: header excessBlobGas {} and calculated excessBlobGas {} do not match",
           headerExcessBlobGas,
           calculatedExcessBlobGas);
+      return false;
+    }
+    long headerBlobGasUsed = header.getBlobGasUsed().orElse(0L);
+    if (headerBlobGasUsed % CancunGasCalculator.BLOB_GAS_PER_BLOB != 0) {
+      LOG.info(
+          "blob gas used must be multiple of GAS_PER_BLOB ({})",
+          CancunGasCalculator.BLOB_GAS_PER_BLOB);
       return false;
     }
     return true;